pyproject.toml
[project]
name = "nexus-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "copilotkit>=0.1.76",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.115.14",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.7",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.7",
    "llama-parse>=0.6.91",
    "mcp[cli]>=1.26.0",
    "mem0ai>=1.0.2",
    "nemoguardrails>=0.20.0",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.16.2",
    "uvicorn>=0.40.0",
]
[tool.uv]
override-dependencies = [
    "fastapi>=0.115.0"
]
